+++ sdk/storage/azblob/CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Features Added
 
 * Added BlobDeleteType to DeleteOptions to allow access to 'Permanent' DeleteType.
+* Added [Set Blob Expiry API](https://learn.microsoft.com/rest/api/storageservices/set-blob-expiry).
+* Added method `ServiceClient()` to the `azblob.Client` type, allowing access to the underlying service client.
 
 ### Breaking Changes